Extend orientation docstring + fix small mistakes
Addressing epic issue #279, added some explanation on the orientation values dip, azimuth and polarity in the docstrings.py.
This was requested by users, as it was unclear how we actually handle these values, azimuth in particular, in GemPy (there are different approaches in general).
Additionally fixed smaller mistakes in the other texts in the docstring.