Fix typos in documentation and update virtual environment source
This commit addresses three changes in the project. In `docs/source/conf.py`, the setting for `pyvista.rcParams['window_size']` was commented out as it was deemed unnecessary for the moment. In `requirements/docs_requirements.txt`, a typo in the 'sphinx-gallery' package name was corrected for proper installation. Moreover, in `scripts/build_docs.sh`, the sourced virtual environment was updated from 'gempy-geotop-pilot' to 'gempy_dependencies' to correctly align with the project's current dependency management setup.