Refine gempy model handling and solver interface
This commit refines aspects of Moureze.py in the examples directory incorporating more data points, increasing refinement, and altering the backend used for processing. It decreases reliance on GPU and changes dtype from 'float64' to 'float32' for performance optimization and to reduce memory usage.
Additionally, references to a plugin 'gempy_plugins' were added to the optional-requirements.txt file to accommodate expanded functionality.
Lastly, in solver_interface.py, the backward propagation on the condition number is commented out, seemingly to avoid unnecessary computation.