Refine Perth basin model and simplify plotting
The refinement parameter for the Perth Basin geological model has been dialed down from 6 to 4 to prevent overfitting. The 'plot_2d' function has been updated to include a default cell_number argument set to 'mid', simplifying the function usage. In addition, a section providing runtime stats, which cluttered the code and was not necessary for functioning, has been removed. These changes make the code cleaner and the model more general.