[ENH] Improve 3D plotting and modify parameters in getting started tutorial
Minor changes were applied to the getting started tutorial, including an adjustment of the pole_vector, specifing that no specific plot resolution is applied, and the use of octrees was noted. Additionally, 3D plotting parameters were modified in both the tutorial and the fault example to fine-tune visualization, specifically the display of surfaces/data in 3D and the display of lithology was enabled in the fault example.